requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argilla==1.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==21.4b2
bs4==0.0.1
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.23
click==8.1.3
clickhouse-connect==0.5.24
cloudpickle==2.2.1
cmake==3.26.3
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
Deprecated==1.2.13
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@d1e04565d3bec8719335b88be9e9b961bf3ec464
docx2txt==0.8
duckdb==0.8.0
effdet==0.4.1
et-xmlfile==1.1.0
fastapi==0.95.2
filelock==3.12.0
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.18.1
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.55.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
iopath==0.1.9
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.176
layoutparser==0.3.4
lit==16.0.5
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.7
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
playwright==1.35.0
portalocker==2.7.0
posthog==3.0.1
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
pyee==9.0.4
Pygments==2.15.1
pypandoc==1.11
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==5.4.1
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sse-starlette==1.6.1
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
threadpoolctl==3.1.0
tiktoken==0.4.0
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.6.8
unstructured-inference==0.4.4
url-normalize==1.4.3
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
Wand==0.6.11
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
wrapt==1.14.1
XlsxWriter==3.1.1
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
